requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
importlib-metadata==4.8.1
Markdown==3.3.4
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
zipp==3.5.0
